Add hot reload button to timeline; move buttons to top.
