+++ lib/ui/icons.dart
@@ -53,6 +53,7 @@ class FlutterIcons {
   static const Icon openTimeline = UrlIcon('/icons/timeline.png');
 
   static const Icon hotRefinal = UrlIcon('/icons/hot-refinal Icon.png');
+  static const Icon hotReload = UrlIcon('/icons/hot-reload.png');
   static const Icon hotRestart = UrlIcon('/icons/hot-restart.png');
 
   static const Icon iconRun = UrlIcon('/icons/refinal Icon_run.png');
